Add IR Analysis for SWP (#189)
Summary:
Adds an IR Analysis step for visualizing a simplified version of the source code with SWP on each loop. This has the following caveats:
1. It doesn't work with WS yet.
2. It won't work with Blackwell operators.

I primarily need this to enable experimenting with a SEV fix that requires fixing the pipeliner. I will add Blackwell support when I resume working on that pipeliner again.
<img width="700" height="447" alt="Screenshot 2025-10-27 at 1 49 20 PM" src="https://github.com/user-attachments/assets/b12c27c5-3c26-4351-8121-8ede1045a0f1" />

Pull Request resolved: https://github.com/meta-pytorch/tritonparse/pull/189

Reviewed By: FindHao, adamomainz

Differential Revision: D85587983

Pulled By: njriasan

fbshipit-source-id: 3145c8924f4179ef1e3d22f9fe664b693916b6da